Refactoring
* Constants splitted
