+++ src/biz.dfch.CS.Abiquo.Client/biz.dfch.CS.Abiquo.Client.csproj
@@ -179,15 +179,17 @@
     <Compile Include="..\AssemblyVersion.cs">
       <Link>Properties\AssemblyVersion.cs</Link>
     </Compile>
+    <Compile Include="Authentication\Constants.cs" />
+    <Compile Include="Communication\AbiquoUrlSuffix.cs" />
     <Compile Include="Factory\AbiquoClientFactory.cs" />
     <Compile Include="Authentication\BasicAuthenticationInformation.cs" />
     <Compile Include="Authentication\IAuthenticationInformation.cs" />
     <Compile Include="BaseAbiquoClient.cs" />
-    <Compile Include="Constants.cs" />
     <Compile Include="ContractClassForBaseAbiquoClient.cs" />
     <Compile Include="LoginResultEnum.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="v1\AbiquoClient.cs" />
+    <Compile Include="v1\AbiquoAcceptHeaders.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="App.config" />
@@ -195,7 +197,6 @@
   </ItemGroup>
   <ItemGroup>
     <Folder Include="Model\" />
-    <Folder Include="v1\Constants\" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it.
